Guard Clause for Rails 7

+++ template.rb
@@ -1,3 +1,10 @@
+if Rails.version >= '7'
+  puts '-----------------'
+  puts "This template doesn't support Rails version 7 and above"
+  puts '-----------------'
+  return
+end
+
 run "if uname | grep -q 'Darwin'; then pgrep spring | xargs kill -9; fi"
 
 # GEMFILE
